Update graphie.py
This version adds a timer to keep it from going too fast and 3 loops.  The first loop runs endlessly.  This is often called a game loop because it causes a game to keep playing.  What do the other 2 loops do?

"input" asks the user for some information and assigns it to a variable.  So, name = "Harry" is really the same as `<name = input ('Name is?" )>`